Browse added

+++ code/autorotate.sh
@@ -1,39 +1,44 @@
 #/bin/bash
 
-cd ~/autorotate/code/
+rootdir=~/GitHub/autorotate/code/
 
+cd $rootdir
 mkdir -p tempstore
 mkdir -p success
 mkdir -p failed
 
-numfil=$(find $1/ -type f | wc -l)
+numfil=$(find $1/ -type f -maxdepth 1| wc -l)
 echo "${numfil} potential files to analyze"
 id=0
 
 for f in "$1"/*
 do
-    
-    ((id++))
-    echo "${id} out of ${numfil}"
-    ext=${f##*.}
-    if [ "$ext" == "tif" ] || [ "$ext" == "tiff" ] || [ "$ext" == "jpg" ] || [ "$ext" == "jpeg" ] || [ "$ext" == "bmp" ] || [ "$ext" == "png" ] ; then
-        echo "Found an image. Analyzing..."
-        ./deskew "$f" > log.txt
 
-        orientation=$(./detectflip tempstore/temp.$ext 2>&1)
+    if [ -f $f ]
+    then
+        ((id++))
+        echo "${id} out of ${numfil}"
+        ext=${f##*.}
+        if [ "$ext" == "tif" ] || [ "$ext" == "tiff" ] || [ "$ext" == "jpg" ] || [ "$ext" == "jpeg" ] || [ "$ext" == "bmp" ] || [ "$ext" == "png" ] ; then
+            echo "Found an image. Analyzing..."
+            ./deskew "$f" > log.txt
 
-        if [ "$orientation" == "upside-down" ]; then
-            ./rotateimg tempstore/temp.$ext
-            mv tempstore/temp.$ext success/$(basename $f)
-        elif [ "$orientation" == "" ]; then
-            echo "Image size too small. Please try using an image of higher resolution."
-            cp $f failed/$(basename $f)
+            orientation=$(./detectflip tempstore/temp.$ext 2>&1)
+
+            if [ "$orientation" == "upside-down" ]; then
+                ./rotateimg tempstore/temp.$ext
+                mv tempstore/temp.$ext success/$(basename $f)
+            elif [ "$orientation" == "" ]; then
+                echo "Image size too small. Please try using an image of higher resolution."
+                cp $f failed/$(basename $f)
+            else
+                mv tempstore/temp.$ext success/$(basename $f)
+            fi
         else
-            mv tempstore/temp.$ext success/$(basename $f)
+            echo "Not an image. Moving on to the next file..."
         fi
-    else
-        echo "Not an image. Moving on to the next file..."
     fi
 done
 
 rm tempstore -rf
+echo "Done. Successful autorotations saved to "$rootdir"success and failures saved to "$rootdir"failed"
